Fix wrong contentInsets for top-level child scroll views
The automaticallyAdjustsScrollViewInsets property makes a container view controller only look at the top level view. In the case of SDCSegmentedViewController, that view was not a scroll view, but a regular UIView. Since SDCSegmentedViewController is itself a container view controller, it should also adjust a top-level scroll view’s insets that it contains.