set segue names from runtime attributes
This allows you to set a value on the Controller called “segueNames” of
type “string” which is a comma separated list of the segue names you
want to put inside the segmented controller.

In other words it replaces the call to `addStoryboardSegments:` with a
property defined on the view controller in the storyboard.