Implement in Swift
Due to difficulties mixing Swift with Objective-C, especially when using Swift-only features, the Swift source is not compiled in the sample project. It is there for reference and can be compiled manually, but there are no unit tests yet. The Swift source is purely experimental and is not yet ready for production. There are no tests for it, and it won't be included in CocoaPods installs.